+++ CHANGELOG.md
@@ -18,6 +18,7 @@ If you had a custom Spring configuration for JGiven you have to change the follo
 ## New Features
 
 * Tags with the same name, but different packages are now correctly distinguished [#242](https://github.com/TNG/JGiven/pull/242) (thanks to ahus1)
+* Scenario states can be marked as required to make scenarios fail quickly and with a clear message if the state hasn't been provided [#255](https://github.com/TNG/JGiven/issues/255)
 
 ### Experimental Android Support